+++ app/src/main/java/com/qinicy/tools/data/api/ICheckPhoneNumApi.java
@@ -4,5 +4,5 @@ package com.qinicy.tools.data.api;
  * Created by qinicy on 16/2/25.
  */
 public interface ICheckPhoneNumApi {
-    void checkPhoneNum(int number,IApiCallBack listener);
+    void checkPhoneNum(Long number,IApiCallBack listener);
 }
