+++ app/src/main/java/com/qinicy/tools/presentation/Interactor/ICheckPhoneNumInteractor.java
@@ -6,5 +6,5 @@ import com.qinicy.tools.data.api.IApiCallBack;
  * Created by qinicy on 16/2/25.
  */
 public interface ICheckPhoneNumInteractor {
-    void checkPhoneNum(int number,IApiCallBack listener);
+    void checkPhoneNum(Long number,IApiCallBack listener);
 }
