+++ app/src/main/java/com/qinicy/tools/presentation/presenter/ICheckPhoneNumPresenter.java
@@ -4,5 +4,6 @@ package com.qinicy.tools.presentation.presenter;
  * Created by qinicy on 16/2/25.
  */
 public interface ICheckPhoneNumPresenter extends IPresenter {
-    void checkPhoneNumInfo(int number);
+
+    void checkPhoneNumInfo(Long number);
 }
